Added a high performance caching. So basically there are multiple ways to cache, one is the invalidate cache, it validates it either manual per route or like in the user routes automatically as soon as something changes. It's useful for a lot of things. It checks and measures errors, hits, misses and more. Was testing it, it reduced the time for db actions from around 250ms to 25-30ms.
I added also redis for the security.middleware.js since there was a problem with rate limiting, so when there was a guest, it limited the requests based on the IP, but somehow there was a problem with the roles, so as soon as I passed the cookie created while login, I could bypass it. So do it now manually.

Added a readme and docs for the caching as well. Changed the docker compose. Now only for dev mode will add it to prod as well.